fixes imported type source rewriting
previously, imported types in extract function were being rewritten as
the full package import paths. this change updates the 'types' package so
that we don't have to modify our vendored version of it, where a change
to the types.TypeString signature now takes a function parameter for
rewriting import paths; we use this new parameter to rewrite the types
how we would like to (via basename instead of path). The 'gcimporter' pkg
also needed updating due to a signature change in the 'types' pkg. All
other packages seem to work. Note that we cannot update the 'loader'
package since it no longer works on cgo.

closes #20