extract func output below current func
this just kind of bugged me, I believe someone else expressed a similar
sentiment that the extracted func should be output below the current enclosed
func rather than at the bottom of the file. this does that. also added an
extra newline preceding the extracted function. this tends to be programmer
preference for whether to put associated functions above or below their caller
(or, for that matter, anywhere remotely close) -- it's rather pedantic to
bicker over, I could put it above if someone so prefers that to be the
default; we could also easily make it configurable.

updated the tests by hand to make sure I didn't fuck anything up. should check
out